pyproject.toml
[project]
name = "meraki-orchestration"
version = "1.0.0"
requires-python = ">=3.9"
dependencies = [
    "meraki>=1.0.0",
    "python-dotenv>=1.0.0",
]

[tool.ruff]
line-length = 127
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = [
    "E501",  # Line too long
    "E722",  # Bare except
    "F841",  # Unused variable
]

[tool.black]
line-length = 127
target-version = ["py313"]
